polish

+++ pod5/manualbench/system/gtgpu-ssd-seq.sh
@@ -8,7 +8,7 @@ die()
 
 cd /data/hasindu/hasindu2008.git/slow5-pod5-bench/pod5 || die "cd fail"
 echo "Copying pod5 file"
-cp /home/hasindu/scratch/hg2_prom_lsk114_5khz/PGXXXX230339_reads.pod5 /data/tmp/PGXXXX230339_reads.pod5 || die "cp fail"
+test -e /data/tmp/PGXXXX230339_reads.pod5 || cp /home/hasindu/scratch/hg2_prom_lsk114_5khz/PGXXXX230339_reads.pod5 /data/tmp/PGXXXX230339_reads.pod5 || die "cp fail"
 
 THREADS=$(nproc)
 
@@ -24,6 +24,6 @@ echo "Iteration $i"
 	./run_seq.sh /data/tmp/PGXXXX230339_reads.pod5 ${THREADS} mmap &> gtgpu-ssd_PGXXXX230339_reads_${THREADS}_mmap_${i}.log
 done
 
-echo "done, remove the file"
-rm /data/tmp/PGXXXX230339_reads.pod5 || die "rm fail"
+echo "done"
+#rm /data/tmp/PGXXXX230339_reads.pod5 || die "rm fail"
